feat: i changed autowired config to construtctor

+++ src/main/java/br/com/meusintoma/modules/calendar/services/CalendarServiceUtils.java
@@ -51,7 +51,7 @@ public class CalendarServiceUtils {
         if (mod == 0)
             return time.truncatedTo(ChronoUnit.MINUTES);
 
-        return time.plusMinutes(slotMinutes - mod).truncatedTo(ChronoUnit.MINUTES);
+        return time.plusMinutes((long) slotMinutes - mod).truncatedTo(ChronoUnit.MINUTES);
     }
 
 }
